refactor: refactoring ContactForm error handling
Creating customHook to management of error states on input fields.
